backend/requirements.txt
# Backend - Flask + MySQL
# Dependencias para ChefCommunity

Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-JWT-Extended==4.7.1
PyMySQL==1.1.1
python-dotenv==1.0.1
bcrypt==4.2.1
cryptography==43.0.1
